Update code.js
Fails test, tried having the function take in the start and end indexes as arguments and would recursively call it, tried another way that would do the search in a while loop, but in any case it still mentions when failing the test that a counterexample to the function is the case of [1]. I would really appreciate some guidance on this. Thanks!